pyproject.toml
[tool.poetry]
name = "11-for-practice"
version = "0.1.0"
description = ""
authors = ["Denis Antropov <de94onisus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.pytest.dependencies]
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
